+++ lib/features/product/domain/entities/product.dart
